+++ CHANGELOG.md
@@ -6,6 +6,30 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+
+## 2024.12.03
+
+### Added
+
+- Designed kernel implementation for CarFuelMonitor component, implemented the transferFrom, newInstance, and clear methods, as well as the methods mentioned in the CarFuelMonitorKernal interface: addFuel, isOutOfFuel, and getFuelLevel.
+
+### Updated
+
+- Changed design to include the kernal implementation for the methods in the CarFuelMonitorKernal interface.
+
+
+
+## 2024.12.01
+
+### Added
+
+- Designed abstract class for CarFuelMonitor component, implemented toString and equals method for this class. Also implemented the methods mentioned in CarFuelMonitorEnhanced: drive, refuel, isLowFuel.
+
+### Updated
+
+- Changed design to include all of the implementations for the methods in the Secondary Interface.
+
+
 ## 2024.10.17
 
 ### Added
@@ -34,9 +58,6 @@ the following form: YYYY.0M.0D.
 - Designed a Car Fuel Monitor component
 - Designed a Home Thermostat component
 - Designed a Stock Tracker component
-
-
-
 
 ```